requirements.txt
streamlit
torch
torchvision
matplotlib
plotly
numpy
pillow
scikit-learn
opencv-python
